rename package pkg/accounting to satellite/accounting
