+++ satellite/accounting/common.go
