+++ satellite/accounting/rollup/common.go
